requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiohttp-retry==2.9.1
aiosignal==1.4.0
altgraph @ file:///AppleInternal/Library/BuildRoots/bcce998f-ff34-11ef-9d34-f2a857e00a32/Library/Caches/com.apple.xbs/Sources/python3/altgraph-0.17.2-py2.py3-none-any.whl
async-timeout==5.0.1
attrs==25.3.0
bleak==1.0.1
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.1.8
contourpy==1.3.0
cycler==0.12.1
decorator==5.2.1
distlib==0.3.9
filelock==3.18.0
fonttools==4.58.5
frozenlist==1.7.0
future @ file:///AppleInternal/Library/BuildRoots/bcce998f-ff34-11ef-9d34-f2a857e00a32/Library/Caches/com.apple.xbs/Sources/python3/future-0.18.2-py3-none-any.whl
geocoder==1.38.1
idna==3.10
importlib_resources==6.5.2
joblib==1.5.1
kiwisolver==1.4.7
macholib @ file:///AppleInternal/Library/BuildRoots/bcce998f-ff34-11ef-9d34-f2a857e00a32/Library/Caches/com.apple.xbs/Sources/python3/macholib-1.15.2-py2.py3-none-any.whl
matplotlib==3.9.4
multidict==6.6.3
numpy==2.0.2
packaging==25.0
pandas==2.3.1
pillow==11.3.0
platformdirs==4.3.8
propcache==0.3.2
PyJWT==2.10.1
pyobjc-core==11.1
pyobjc-framework-Cocoa==11.1
pyobjc-framework-CoreBluetooth==11.1
pyobjc-framework-libdispatch==11.1
pyparsing==3.2.3
pyserial==3.5
python-dateutil==2.9.0.post0
pytz==2025.2
ratelim==0.1.6
requests==2.32.4
scikit-learn==1.6.1
scipy==1.13.1
six @ file:///AppleInternal/Library/BuildRoots/bcce998f-ff34-11ef-9d34-f2a857e00a32/Library/Caches/com.apple.xbs/Sources/python3/six-1.15.0-py2.py3-none-any.whl
threadpoolctl==3.6.0
twilio==9.6.5
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
virtualenv==20.31.2
yarl==1.20.1
zipp==3.23.0
